Setup GerkeLab Default GitHub Labels

Removes the default GitHub labels and installs the GerkeLab labels. A small wrapper around usethis::use_github_labels().

use_grk_github_labels(..., labels = grk_github_labels(),
  colours = grk_github_colours(),
  descriptions = grk_github_descriptions(), delete_default = TRUE)

grk_github_labels()

grk_github_colours()

grk_github_descriptions()

Arguments

...
Arguments passed on to usethis::use_github_labels

repo_spec

Optional repository specification (owner/repo) if you don't want to target the current project.

labels

A character vector giving labels to add.

rename

A named vector with names giving old names and values giving new names.

colours

Named character vectors giving hexadecimal colours (like e02a2a) and longer descriptions. The names should match label names, and anything unmatched will be left unchanged. If you create a new label, and don't supply colours, it will be given a random colour.

descriptions

Named character vectors giving hexadecimal colours (like e02a2a) and longer descriptions. The names should match label names, and anything unmatched will be left unchanged. If you create a new label, and don't supply colours, it will be given a random colour.

delete_default

If TRUE, removes GitHub default labels that do not appear in the labels vector and that do not have associated issues.

auth_token

GitHub personal access token (PAT).

host

GitHub API host to use. Override with the endpoint-root for your GitHub enterprise instance, for example, "https://github.hostname.com/api/v3".

labels
A character vector giving labels to add.

colours
Named character vectors giving hexadecimal colours (like e02a2a) and longer descriptions. The names should match label names, and anything unmatched will be left unchanged. If you create a new label, and don't supply colours, it will be given a random colour.

descriptions
Named character vectors giving hexadecimal colours (like e02a2a) and longer descriptions. The names should match label names, and anything unmatched will be left unchanged. If you create a new label, and don't supply colours, it will be given a random colour.

delete_default
If TRUE, removes GitHub default labels that do not appear in the labels vector and that do not have associated issues.

Functions

grk_github_labels: Default grkmisc GitHub labels.

grk_github_colours: Default grkmisc GitHub colours.

grk_github_descriptions: Default grkmisc GitHub descriptions.