Create a default package skeleton

Create a default package skeleton

use_grk_starter_package(path, github = TRUE, github_org = NULL,
  github_private = FALSE, open = TRUE,
  title = "What the Package Does (One Line, Title Case)",
  description = "What the package does (one paragraph)")

Arguments

path
Path (and thus package name) where the package will be stored

github
Should a github repository be created?

github_org
The organization where the repo should be created. If unspecified (or NULL), the default is to create the package in your personal account.

github_private
Should the GitHub repo be private?

open
Should the package be opened? If TRUE, tries to open in RStudio and falls back to changing the working directory.

title
Package title: "What the Package Does (One Line, Title Case)"

description
Package description: "What the package does (one paragraph)"

Default Package Description

See usethis::use_description() for information about setting your default package DESCRIPTION file fields. The help file there provides an example of how to set "usethis.description" in your .Rprofile.